+++ Cardboard.Net.Core/Entities/IDeletable.cs
@@ -1,0 +1,9 @@
+namespace Cardboard;
+
+public interface IDeletable
+{
+    /// <summary>
+    /// Deletes this object and all its children
+    /// </summary>
+    Task DeleteAsync();
+}
