+++ Cardboard.Net.Core/Entities/IMisskeyEntity.cs
@@ -1,0 +1,6 @@
+namespace Cardboard;
+
+public interface IMisskeyEntity : IEquatable<string>
+{
+    string Id { get; }
+}
