+++ Cardboard.Net.Core/Entities/Notes/AcceptanceType.cs
@@ -2,7 +2,7 @@ using System.Runtime.Serialization;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace Cardboard.Net.Entities.Notes;
+namespace Cardboard.Notes;
 
 /// <summary>
 /// Reaction acceptance
