+++ Cardboard.Net.Core/Entities/Roles/RoleType.cs
