+++ Cardboard.Net/Entities/Users/Achievement.cs
@@ -1,3 +1,4 @@
+using Cardboard.Users;
 using Newtonsoft.Json;
 
 namespace Cardboard.Net.Entities.Users;
